+++ app/src/main/java/com/apap/forecasty/domain/model/Temperature.kt
@@ -5,10 +5,10 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class Temperature(
-    val day: Double,
-    val min: Double,
-    val max: Double,
-    val night: Double,
-    val evening: Double,
-    val morning: Double,
+    val day: Float,
+    val min: Float,
+    val max: Float,
+    val night: Float,
+    val evening: Float,
+    val morning: Float,
 ) : Parcelable
